Rename utility script file, move Firebase handling to Firebase class
